Move InPlaceSeed out of private mod
